@@ -3,10 +3,12 @@ import 'package:rapidoreach/RapidoReach.dart';
 
 void main() {
   WidgetsFlutterBinding.ensureInitialized();
-  runApp(MyApp());
+  runApp(const MyApp());
 }
 
 class MyApp extends StatefulWidget {
+  const MyApp({Key? key}) : super(key: key);
+
   @override
   _MyAppState createState() => _MyAppState();
 }
@@ -15,7 +17,7 @@ class _MyAppState extends State<MyApp> {
   @override
   void initState() {
     RapidoReach.instance
-        .init(apiToken: 'YOUR_APP_API_KEY', userId: 'ANDROID_TEST_ID');
+        .init(apiToken: 'e8ad9ecc9dd75f140cd662565da0954e', userId: '22992929');
     RapidoReach.instance.setOnRewardListener(onRapidoReachReward);
     RapidoReach.instance.setRewardCenterClosed(onRapidoReachRewardCenterClosed);
     RapidoReach.instance.setRewardCenterOpened(onRapidoReachRewardCenterOpened);
@@ -23,24 +25,24 @@ class _MyAppState extends State<MyApp> {
         .setSurveyAvaiableListener(onRapidoReachSurveyAvailable);
     RapidoReach.instance.setNavBarText(text: 'RapidoReach');
     RapidoReach.instance.setNavBarColor(color: '#211548');
-    RapidoReach.instance.setNavBarTextColor(text_color: '#FFFFFF');
+    RapidoReach.instance.setNavBarTextColor(textColor: '#FFFFFF');
     super.initState();
   }
 
   void onRapidoReachReward(int? quantity) {
-    print('ROR: $quantity');
+    debugPrint('ROR: $quantity');
   }
 
   void onRapidoReachSurveyAvailable(int? survey) {
-    print('ROR: $survey');
+    debugPrint('ROR: $survey');
   }
 
   void onRapidoReachRewardCenterClosed() {
-    print('ROR: closed');
+    debugPrint('ROR: closed');
   }
 
   void onRapidoReachRewardCenterOpened() {
-    print('ROR: opened');
+    debugPrint('ROR: opened');
   }
 
   @override
@@ -54,12 +56,12 @@ class _MyAppState extends State<MyApp> {
             child: Column(
           mainAxisAlignment: MainAxisAlignment.center,
           children: <Widget>[
-            RaisedButton(
-              child: Text("Launch RapidoReach"),
+            ElevatedButton(
+              child: const Text("Launch RapidoReach"),
               onPressed: () => RapidoReach.instance.show(),
             ),
-            RaisedButton(
-              child: Text("Launch RapidoReach Placement"),
+            ElevatedButton(
+              child: const Text("Launch RapidoReach Placement"),
               onPressed: () => RapidoReach.instance.show(),
             )
           ],
